Suggestion for documentation update
Adding a note addressed to plugin/theme authors that are using Carbon-Fields as a plugin, and not requiring it from their theme/plugin. In that case a fallback functions needs to be defined, in case the user deactivates the Carbon-Fields plugin. Otherwise the entire site will break by the usage of undefined functions.